// src/app/models/CategoryModels.ts
export interface ICategory {
  id: string;
  name: string;
  description?: any;
  imageUrl: string;
  createdOn: string;
  baseCategoryId?: string | null;
  baseCategory?: string | null;
  code?: string | null;
  iconUrl?: null | string;
  isTopCategory?: boolean;
  orderingNumber?: null | number;
}

// Generated by https://quicktype.io

export interface CategoryResponse {
  message: string;
  status: string;
  data: ICategory[];
}